requirements.txt
openai
langchain_community
streamlit
langgraph
python-dotenv
google-search-results